+++ tests/test.rs
@@ -320,13 +320,14 @@ fn comment() {
     assert_eq!(changelog["0.2.0"].notes, "<!--\n# 0.2.0-->");
     assert_eq!(changelog["0.1.0"].notes, "<!--\n# 0.1.0-->");
 
+    // https://pandoc.org/try/?params=%7B%22text%22%3A%22%23+0.4.0%5Cna+%3C%21--%5Cn%23+0.3.0+--%3E+%3C%21--%5Cn%23+0.2.0+--%3E%5Cn%23+0.1.0%5Cn+%3C%21--+%3C%21--%5Cn%23+0.1.0+%3C%21--+--%3E%5Cna%5Cn%23+0.0.2+--%3E%5Cn%23+0.0.1+%3C%21--+--%3E%5Cn%23+0.0.0+%3C%21--%5Cna%5Cn--%3E%5Cn%3C%21--%5Cn%23+0.0.0+--%3E%22%2C%22to%22%3A%22html5%22%2C%22from%22%3A%22commonmark%22%2C%22standalone%22%3Afalse%2C%22embed-resources%22%3Afalse%2C%22table-of-contents%22%3Afalse%2C%22number-sections%22%3Afalse%2C%22citeproc%22%3Afalse%2C%22html-math-method%22%3A%22plain%22%2C%22wrap%22%3A%22auto%22%2C%22highlight-style%22%3Anull%2C%22files%22%3A%7B%7D%2C%22template%22%3Anull%7D
     let changelog = "\
-# 0.2.0
-<!--
-# 0.2.0 --> <!--
+# 0.4.0
+a <!--
+# 0.3.0 --> <!--
 # 0.2.0 -->
 # 0.1.0
-<!--
+ <!-- <!--
 # 0.1.0 <!-- -->
 a
 # 0.0.2 -->
@@ -338,17 +339,20 @@ a
 # 0.0.0 -->
 ";
     let changelog = parse(changelog).unwrap();
-    assert_eq!(changelog.len(), 5);
-    assert_eq!(changelog["0.2.0"].title, "0.2.0");
-    assert_eq!(changelog["0.2.0"].notes, "<!--\n# 0.2.0 --> <!--\n# 0.2.0 -->");
+    assert_eq!(changelog.len(), 7);
+    assert_eq!(changelog["0.4.0"].title, "0.4.0");
+    assert_eq!(changelog["0.4.0"].notes, "a <!--");
+    assert_eq!(changelog["0.3.0"].title, "0.3.0 --> <!--");
+    assert_eq!(changelog["0.3.0"].notes, "");
+    assert_eq!(changelog["0.2.0"].title, "0.2.0 -->");
+    assert_eq!(changelog["0.2.0"].notes, "");
     assert_eq!(changelog["0.1.0"].title, "0.1.0");
-    assert_eq!(changelog["0.1.0"].notes, "<!--\n# 0.1.0 <!-- -->\na");
+    assert_eq!(changelog["0.1.0"].notes, " <!-- <!--\n# 0.1.0 <!-- -->\na");
     assert_eq!(changelog["0.0.2"].title, "0.0.2 -->");
     assert_eq!(changelog["0.0.2"].notes, "");
     // TODO: option like title_no_link to remove comment from title
     assert_eq!(changelog["0.0.1"].title, "0.0.1 <!-- -->");
     assert_eq!(changelog["0.0.1"].notes, "");
-    // https://pandoc.org/try/?params=%7B%22text%22%3A%22%23+0.0.0+%3C%21--%5Cna%5Cn--%3E%5Cn%3C%21--%5Cn%23+0.0.0+--%3E%22%2C%22to%22%3A%22html5%22%2C%22from%22%3A%22commonmark%22%2C%22standalone%22%3Afalse%2C%22embed-resources%22%3Afalse%2C%22table-of-contents%22%3Afalse%2C%22number-sections%22%3Afalse%2C%22citeproc%22%3Afalse%2C%22html-math-method%22%3A%22plain%22%2C%22wrap%22%3A%22auto%22%2C%22highlight-style%22%3Anull%2C%22files%22%3A%7B%7D%2C%22template%22%3Anull%7D
     assert_eq!(changelog["0.0.0"].title, "0.0.0 <!--");
     assert_eq!(changelog["0.0.0"].notes, "a\n-->\n<!--\n# 0.0.0 -->");
 }
@@ -473,7 +477,7 @@ fn cargo() {
 #[test]
 fn fuzz() {
     let tests: &[(&str, Result<usize, &str>)] =
-        &[("1115.8.8 '9.\n-\n\u{c}\"----\u{19}\u{1f}<!--.4\n## 444.444.4\r\r \u{b}---->", Ok(1))];
+        &[("1115.8.8 '9.\n-\n\u{c}\"----\u{19}\u{1f}<!--.4\n## 444.444.4\r\r \u{b}---->", Ok(2))];
     for &(test, expected_len) in tests {
         let res = parse(test);
         match expected_len {